requirements.txt
docopt==0.6.2
glob2==0.7
hydra-core==0.11.3
ipython==7.13.0
json5==0.9.5
jupyterlab==2.1.3
matplotlib==3.2.1
more-itertools==8.3.0
notebook==6.0.3
numpy==1.18.1
omegaconf==1.4.1
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==7.0.0
pytest==5.4.2
pytorch-lightning==0.7.6
scikit-learn==0.23.1
scipy==1.4.1
tqdm==4.42.1
